SDR-331 test: updateImageMetaData 메서드 테스트 생성 (pass X)
